Up-to-date checking: Don't re-format unnecessarily
Gradle's UP-TO-DATE checking is based on the up-to-dateness of the
'inputs' and 'outputs' of a task. When we've changed a java source
file, we would like to run the formatter again to ensure correct style.
Gradle is able to detect that an input has changed (because of _our_
editing), but it will also (in case the formatter corrected style
errors) detect a changed input for a following up run of our task (because
of the _formatter's_ editing).
We introduced our own build cache with file state information that takes
this into account. If a format task changes a file, it won't re-format
this file again on the next run.
However, we do _not_ in any way influence Gradle's caching
mechanism. Other tasks, e.g. javaCompile, _will_ run again if the
formatter has changed anything. This was a deliberate decision. The
formatter _should_ make only whitespace changes (-> same bytecode), but
who knows... bugs are everywhere ;)