Make default tasks' inputs resettable
Previously the inputs where assigned in afterEvaluate{...}. Therefore it
was impossible to reset the inputs in the build script.

Now the inputs are assigned immediately when the plugin is applied (with
a lazy and live FileTree). In a build script it is now possible to use
the setSource(..) method to completely discard the default inputs.