Version specific formatter factory; simplified API
With this commit we add support for building version specific formatter
factories. This should make it easier to support new versions of
google-java-format in the future. An initial factory for version "1.0"
is also part of this commit.

Also, we added another factory method to the API that only
allows to specify the style (GOOGLE or AOSP) and no other formatter
option. We use sensible defaults for the rest.