Make default inputs configurable via the extension
The extension object now has the same functions as a task
object (setSource, source, include, exclude). The filtered set of files
is passed on to the tasks before they are executed. The task only adds
this set of files to it's inputs, if no inputs were assigned to this
task before. The additional set of files is further filtered according
to the task's exclude/include patterns (if existent).

The default inputs now apply to _all_ tasks, not only to the default
tasks `googleJavaFormat` and `verifyGoogleJavaFormat`. However, since a
custom task should have some sources set (otherwise it would do
nothing), this change doesn't affect them.