Allow filtering tasks' inputs via system properties
An additional include filter can now be applied on the command line at
execution time by setting the "<taskName>.include" system property.

The property's value is a comma separated list of Ant-style include
patterns. In addition to everything that was defined statically in
'build.gradle', an input file has to match one of these patterns in
order to be included as input by the corresponding task.